% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildParentTree.R
\name{buildParentTree}
\alias{buildParentTree}
\title{Build parent clustering tree}
\usage{
buildParentTree(
  object,
  key = "CHOIR",
  distance_approx = TRUE,
  normalization_method = "none",
  reduction_method = NULL,
  reduction_params = list(),
  n_var_features = NULL,
  batch_correction_method = "none",
  batch_correction_params = list(),
  batch_labels = NULL,
  neighbor_params = list(),
  cluster_params = list(algorithm = 1, group.singletons = TRUE),
  use_assay = NULL,
  use_slot = NULL,
  ArchR_matrix = NULL,
  ArchR_depthcol = NULL,
  reduction = NULL,
  var_features = NULL,
  atac = FALSE,
  n_cores = NULL,
  random_seed = 1,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class 'Seurat' (any version, but v5 is recommended),
'SingleCellExperiment', or 'ArchRProject'.}

\item{key}{The name under which CHOIR-related data for this run is stored in
the object. Defaults to 'CHOIR'.}

\item{distance_approx}{A boolean value indicating whether or not to use
approximate distance calculations. Default = TRUE will use centroid-based
distances.}

\item{normalization_method}{A character string or vector indicating which
normalization method to use. In general, input data should be supplied to
CHOIR after normalization, except in cases when the user wishes to use
\code{Seurat::SCTransform()} normalization. Permitted values are 'none' or
'SCTransform'. Defaults to 'none'.}

\item{reduction_method}{A character string or vector indicating which
dimensionality reduction method to use. Permitted values are 'PCA' for
principal component analysis, 'LSI' for latent semantic indexing, and
'IterativeLSI' for iterative latent semantic indexing. Default = \code{NULL}
will specify a method based on the input data type.}

\item{reduction_params}{A list of additional parameters to be passed to
the selected dimensionality reduction method.}

\item{n_var_features}{A numerical value indicating how many variable
features to identify. Default = \code{NULL} will use 2000 features, or 25000
features for ATAC-seq data.}

\item{batch_correction_method}{A character string or vector indicating which
batch correction method to use. Permitted values are 'Harmony' and
'none'. Defaults to 'none'.}

\item{batch_correction_params}{A list of additional parameters to be passed
to the selected batch correction method for each iteration. Only applicable
when 'batch_correction_method' = 'Harmony'.}

\item{batch_labels}{If applying batch correction, the name of the column
containing the batch labels. Defaults to \code{NULL}.}

\item{neighbor_params}{A list of additional parameters to be passed to
\code{Seurat::FindNeighbors()} (or, in the case of multi-modal data for
Seurat or SingleCellExperiment objects,
\code{Seurat::FindMultiModalNeighbors()}).}

\item{cluster_params}{A list of additional parameters to be passed to
Seurat::FindClusters() for clustering at each level of the tree. Note that if
'group.singletons' is set to TRUE, clusters are relabeled such that each
singleton constitutes its own cluster.}

\item{use_assay}{For Seurat or SingleCellExperiment objects, a character
string or vector indicating the assay(s) to use in the provided object.
Default = \code{NULL} will choose the current active assay for Seurat objects
and the \code{log_counts} assay for SingleCellExperiment objects.}

\item{use_slot}{For Seurat objects, a character string or vector indicating
the slot(s) (Seurat v4) or layer(s) (Seurat v5) to use in the provided object.
Default = \code{NULL} will choose a slot/layer based on the selected assay.
If a non-standard assay is provided, do not leave \code{use_slot} as \code{NULL}.}

\item{ArchR_matrix}{For ArchR objects, a character string or vector
indicating which matri(ces) to use in the provided object. Default =
\code{NULL} will use the 'TileMatrix' for ATAC-seq data or the
'GeneExpressionMatrix' for RNA-seq data.}

\item{ArchR_depthcol}{For ArchR objects, a character string or vector
indicating which column to use for correlation with sequencing depth.
Default = \code{NULL} will use the 'nFrags' column for ATAC-seq data or the
'Gex_nUMI' for RNA-seq data.}

\item{reduction}{An optional matrix of dimensionality reduction cell
embeddings to be used for subsequent clustering steps. Defaults to
\code{NULL}, whereby dimensionality reduction(s) will be calculated using
method specified by 'reduction_method' as part of the \code{buildTree()}
function.}

\item{var_features}{An optional character vector of variable features to be
used for subsequent clustering steps. Defaults to \code{NULL}, whereby
variable features will be calculated as part of the \code{buildTree()}
function.}

\item{atac}{A boolean value or vector indicating whether the provided data is
ATAC-seq data. Defaults to \code{FALSE}.}

\item{n_cores}{A numeric value indicating the number of cores to use for
parallelization. Default = \code{NULL} will use the number of available cores
minus 2.}

\item{random_seed}{A numeric value indicating the random seed to be used.}

\item{verbose}{A boolean value indicating whether to use verbose output
during the execution of this function. Can be set to \code{FALSE} for a
cleaner output.}
}
\value{
Returns the object with the following added data stored under the
provided key: \describe{
\item{reduction}{Cell embeddings for calculated dimensionality reduction}
\item{var_features}{Variable features for calculated dimensionality reduction}
\item{cell_IDs}{Cell IDs belonging to parent tree}
\item{graph}{Nearest neighbor and shared nearest neighbor adjacency matrices}
\item{clusters}{Parent hierarchical cluster tree}
\item{parameters}{Record of parameter values used}
}
}
\description{
This function constructs a hierarchical clustering tree starting from a single
cluster encompassing all cells. A parent tree is constructed, from which
subtrees can be generated with subsequent steps outside of this function.
}
\details{
For multi-modal data, optionally supply parameter inputs as vectors/lists
that sequentially specify the value for each modality.
}
