% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHOIR-package.R
\docType{package}
\name{CHOIR-package}
\alias{CHOIR-package}
\title{CHOIR: CHOIR - Clustering Hierachy Optimization by Iterative Random forests}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

CHOIR is a clustering algorithm for single-cell sequencing data. CHOIR applies a framework of permutation tests and random forest classifiers across a hierarchical clustering tree to statistically identify clusters that represent distinct populations.
}
\author{
\strong{Maintainer}: Cathrine Petersen \email{cathrine.petersen@gladstone.ucsf.edu} (\href{https://orcid.org/0000-0002-5821-9828}{ORCID})

}
\keyword{internal}
