% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{CHOIRpalette}
\alias{CHOIRpalette}
\title{Generate color palette}
\usage{
CHOIRpalette(n)
}
\arguments{
\item{n}{Number of colors}
}
\value{
Returns a vector of n hex values.
}
\description{
Generate a color palette based on input number. Hex values are standard up to
n = 100, but for larger values of n, they are generated using
\code{Polychrome::createPalette}.
}
