% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{getRecords}
\alias{getRecords}
\title{Retrieve all CHOIR metadata}
\usage{
getRecords(object, key = "CHOIR")
}
\arguments{
\item{object}{An object of class \code{Seurat}, \code{SingleCellExperiment},
or \code{ArchRProject} that has undergone CHOIR clustering.}

\item{key}{The name under which CHOIR-related data for this run is stored in
the object. Defaults to “CHOIR”.}
}
\value{
Returns the data stored under the provided key.
}
\description{
Retrieve all stored CHOIR data from object
}
