% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{plotCHOIR}
\alias{plotCHOIR}
\title{Dimensional reduction plot with CHOIR clusters labeled}
\usage{
plotCHOIR(
  object,
  key = "CHOIR",
  reduction = NULL,
  cells = NULL,
  legend = TRUE,
  accuracy_scores = FALSE,
  plot_nearest = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class 'Seurat', 'SingleCellExperiment', or
'ArchRProject' that has undergone CHOIR clustering.}

\item{key}{The name under which CHOIR-related data for this run is retrieved
from the object. Defaults to 'CHOIR'.}

\item{reduction}{A character vector indicating which CHOIR subtree
dimensionality reduction to plot (e.g., 'P0_reduction_UMAP',
'P1_reduction_UMAP'). Default = \code{NULL} will plot the UMAP of the
'P0_reduction_UMAP' generated by CHOIR.}

\item{cells}{Vector of cell names to plot. Default is all cells.}

\item{legend}{A boolean value indicating whether to include a legend. Defaults to
\code{TRUE}.}

\item{accuracy_scores}{A boolean value indicating whether to overlay pairwise
cluster comparison accuracy scores onto UMAP. Default = \code{TRUE}.}

\item{plot_nearest}{A boolean value indicating whether to only plot the
accuracy score for each cluster's nearest neighboring cluster. Only relevant
if \code{accuracy_scores} is \code{TRUE}. Defaults to \code{TRUE}.}

\item{...}{Extra parameters passed to \code{Seurat::DimPlot()}}
}
\value{
A ggplot object.
}
\description{
Dimensional reduction plot with CHOIR clusters labeled
}
