% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCountSplit.R
\name{runCountSplit}
\alias{runCountSplit}
\title{Run countsplitting}
\usage{
runCountSplit(
  object,
  key = "CHOIR",
  use_assay = NULL,
  use_slot = NULL,
  ArchR_matrix = NULL,
  countsplit_suffix = c("_1", "_2"),
  countsplit_params = list(),
  normalization_method = "log",
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class 'Seurat', 'SingleCellExperiment', or
'ArchRProject'.}

\item{key}{The name under which CHOIR-related data for this run is stored in
the object. Defaults to 'CHOIR'.}

\item{use_assay}{For Seurat or SingleCellExperiment objects, a character
string or vector indicating the assay(s) to use in the provided object.
Default = \code{NULL} will choose the current active assay for Seurat objects
and the \code{counts} assay for SingleCellExperiment objects.}

\item{use_slot}{For Seurat objects, a character string or vector indicating
the layers(s) — previously known as slot(s) — to use in the provided object.
Default = \code{NULL} will use the 'counts' slot.}

\item{ArchR_matrix}{For ArchR objects, a character string or vector
indicating which matri(ces) to use in the provided object. Default =
\code{NULL} will use the 'TileMatrix' for ATAC-seq data or the
'GeneExpressionMatrix' for RNA-seq data.}

\item{countsplit_suffix}{A character vector indicating the suffixes
that distinguish the two countsplit matrices to be used. Suffixes are
appended onto input string/vector for \code{use_slot} for Seurat objects,
\code{use_assay} for SingleCellExperiment objects, or \code{ArchR_matrix} for
ArchR objects. Defaults to suffixes "_1" and "_2".}

\item{countsplit_params}{A list of additional parameters to be passed to
\code{countsplit::countsplit()}.}

\item{normalization_method}{A character string or vector indicating which
normalization method to apply after countsplitting. Permitted values are
'none', 'log', or 'tfidf'. Defaults to 'log'.}

\item{verbose}{A boolean value indicating whether to use verbose output
during the execution of this function. Can be set to \code{FALSE} for a
cleaner output.}
}
\value{
Returns the object including the newly generated countsplit matrices
and the following added data stored under the provided key: \describe{
\item{parameters}{Record of parameter values used}
}
}
\description{
Fetches specified count matri(ces) from the provided input object, generates
countsplit matrices using function \code{countsplit::countsplit()} from A.
Neufeld, and stores these matrices back in the object with suffixes
provided by 'countsplit_suffix'.
}
