% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeUtils.R
\name{inferTree}
\alias{inferTree}
\title{Infer clustering tree from pre-generated clusters}
\usage{
inferTree(cluster_labels, reduction = NULL, verbose = TRUE)
}
\arguments{
\item{cluster_labels}{A named vector of cluster IDs. Names must correspond
to cell IDs.}

\item{reduction}{A matrix of dimensionality reduction cell
embeddings to be used for distance calculations.}

\item{verbose}{A Boolean value indicating whether to use verbose output
during the execution of this function. Can be set to \code{FALSE} for a
cleaner output.}
}
\value{
A clustering tree as a dataframe.
}
\description{
Generate clustering tree from provided, pre-generated clusters. Provide a set
of cluster labels and a dimensionality reduction.
}
