% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{plotCHOIR}
\alias{plotCHOIR}
\title{Dimensional reduction plot with CHOIR clusters labeled}
\usage{
plotCHOIR(
  object,
  key = "CHOIR",
  reduction = NULL,
  group_by = NULL,
  cells = NULL,
  highlight_cells = NULL,
  label = TRUE,
  legend = TRUE,
  accuracy_scores = FALSE,
  plot_nearest = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class 'Seurat', 'SingleCellExperiment', or
'ArchRProject' that has undergone CHOIR clustering.}

\item{key}{The name under which CHOIR-related data for this run is retrieved
from the object. Defaults to 'CHOIR'.}

\item{reduction}{A character vector indicating which CHOIR subtree
dimensionality reduction to plot (e.g., 'P0_reduction_UMAP',
'P1_reduction_UMAP'). Default = \code{NULL} will plot the UMAP of the
'P0_reduction_UMAP' generated by CHOIR.}

\item{group_by}{A character string indicating which metadata column to color
cells by. Default = \code{NULL} will color cells according to the clusters
identified by CHOIR.}

\item{cells}{Vector of cell names to plot. Default is all cells.}

\item{highlight_cells}{Vector of cell names to highlight. Default =
\code{NULL}.}

\item{label}{A boolean value indicating whether to label groups.
Defaults to \code{TRUE}.}

\item{legend}{A boolean value indicating whether to include a legend.
Defaults to \code{TRUE}.}

\item{accuracy_scores}{A boolean value indicating whether to overlay pairwise
cluster comparison accuracy scores onto UMAP. Default = \code{TRUE}.}

\item{plot_nearest}{A boolean value indicating whether to only plot the
accuracy score for each cluster's nearest neighboring cluster. Only relevant
if \code{accuracy_scores} is \code{TRUE}. Defaults to \code{TRUE}.}

\item{...}{Extra parameters passed to \code{Seurat::DimPlot()}.}
}
\value{
A ggplot object.
}
\description{
Generate a dimensionality reduction embedding plot colored (by default)
according to the clusters identified by CHOIR. Additionally permits plotting
mean prediction accuracy scores for pairs of clusters as an overlay.
}
